refactor: :truck: Moved aircraft to a packag, reorder and add missing annotations
