+++ src/main/java/acme/entities/aircrafts/AircraftStatus.java
@@ -1,5 +1,5 @@
 
-package acme.entities;
+package acme.entities.aircrafts;
 
 public enum AircraftStatus {
 	ACTIVE, UNDER_MAINTENANCE
